Refactor

+++ server/src/main/java/com/weshare/repository/UserRepository.java
@@ -6,10 +6,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
     Optional<User> findUserByName(String name);
-    Optional<User> findUserById(Long id);
+    Optional<User> findUserById(Integer id);
 
     Optional<List<User>> findUsersByNameIn(List<String> names);
